+++ package.json
@@ -21,11 +21,13 @@
   "scripts": {
     "check": "npm install && npm outdated",
     "update": "ncu -u && npm update && npm install",
-    "updateDependencies": "npm run check && npm run update",
+    "update dependencies": "npm run check && npm run update",
     "lint": "eslint src/**.ts",
-    "watch": "npm run build && npm link && nodemon",
+    "jlint": "eslint homebridge-ui/public/**.mjs",
+    "watch": "npm run build && npm run plugin-ui && npm link && nodemon",
+    "plugin-ui": "rsync ./src/homebridge-ui/public/index.html ./dist/homebridge-ui/public/",
     "build": "rimraf ./dist && tsc",
-    "prepublishOnly": "npm run lint && npm run build",
+    "prepublishOnly": "npm run lint && npm run build && npm run plugin-ui ",
     "postpublish": "npm run clean",
     "clean": "rimraf ./dist",
     "test": "eslint src/**.ts"
@@ -57,12 +59,12 @@
     "ir"
   ],
   "dependencies": {
-    "@homebridge/plugin-ui-utils": "^1.0.0",
+    "@homebridge/plugin-ui-utils": "^1.0.1",
     "async-mqtt": "^2.6.3",
     "fakegato-history": "^0.6.4",
     "homebridge-lib": "^6.7.3",
     "rxjs": "^7.8.1",
-    "undici": "^6.2.1"
+    "undici": "^6.4.0"
   },
   "optionalDependencies": {
     "node-switchbot": "^1.10.0",
@@ -70,12 +72,12 @@
     "@abandonware/bluetooth-hci-socket": "^0.5.3-10"
   },
   "devDependencies": {
-    "@types/node": "^20.10.6",
-    "@typescript-eslint/eslint-plugin": "^6.17.0",
-    "@typescript-eslint/parser": "^6.17.0",
+    "@types/node": "^20.11.5",
+    "@typescript-eslint/eslint-plugin": "^6.19.0",
+    "@typescript-eslint/parser": "^6.19.0",
     "eslint": "^8.56.0",
     "homebridge": "^1.7.0",
-    "nodemon": "^3.0.2",
+    "nodemon": "^3.0.3",
     "npm-check-updates": "^16.14.12",
     "rimraf": "^5.0.5",
     "ts-node": "^10.9.2",
