+++ src/homebridge-ui/public/index.html
